Add primitive printk-based logging
Support the `log` crate's log macros, using Zephyr's printk mechanism.
There is no filter, or deferral of messages, and it pulls in both Rust
and Zephyr's string formatting, but does allow logging to work directly.

This is intended as a stopgap until better logging support comes in, but
is also useful in its own right for builds where Zephyr logging is
disabled.

Signed-off-by: David Brown <david.brown@linaro.org>