zephyr-build: YAML-based augment code
Instead of hardcoding all of the augments in the code, put these
augments into a data structure.  Load the actual rules from a yaml file
(with a future ability to extend this with per-app or per-module defined
augments.

Convert all of the existing hardcoded augment rules into the initial
augment file.