zephyr: device: uart: Implement a blocking transmit
A blocking write allows us to do the write without needing an
intermediate copy of the data.  The thread's buffer is directly copied
into the uart's fifo.

To make this safe, the write call must block until all of the data has
been placed in the fifo.

The uart.h irq interface is a little quirky.  Namely, the uart fifo can
only be accessed from the irq handler.  To make this possible, the
drivers will implement usually a soft interrupt to the irq handler, for
the initial transmission.

There is also an async interface in uart.h, but numerous devices don't
support it.

Signed-off-by: David Brown <david.brown@linaro.org>